// apps/ed-tech.co/app/NX/types.d.ts
export interface I_Nav {
    navItems: I_NavNode[];
    mode?: 'mobile' | 'desktop';
    frontmatter?: T_Frontmatter;
}
// ...
export type T_Frontmatter = {
    title?: string;
    description?: string;
    slug?: string;
    tags?: string;
    icon?: string;
    order?: number;
    image?: string;
    layout?: string;
    hideInNav?: boolean | string;
};

export type T_Markdown = {
    id: string;
    published: boolean;
    createdAt: string;
    updatedAt: string;
    frontmatter?: T_Frontmatter;
};

export interface I_Header {
    title: string;
    description?: string;
    icon?: string;
}

export interface I_NavNode {
    title?: string;
    slug?: string;
    path?: string;
    children?: I_NavNode[];
    icon?: string;
}